Use versions suitable for `-Z minimum-versions`
Unfortunately, even the latest `compiletest_rs` cannot be built with
`-Z minimum-versions`, but since it is actually a dev-dependency, we can
allow ourselves not to care about keeping it compatible with crates from
ecosystem.

Also rename the "nightly" Cargo feature to "test-nightly" to better
reflect its purpose and to prevent users from employing it.